apps
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/gwidgets@407 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/gWidgetsWWW2/R/gpanedgroup.R
@@ -56,7 +56,7 @@ gpanedgroup <- function(horizontal=TRUE, container=NULL, ...,
 GPanedGroup <- setRefClass("GPanedGroup",
                            contains="ExtContainer",
                            fields=list(
-                             child_ct = "integer",
+                             child_ct = "numeric",
                              horizontal="logical"
                              ),
                            method=list(
